requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3
psycopg2-binary==2.9.7
python-decouple==3.8
Pillow==10.0.1
celery==5.3.4
django-celery-beat==2.5.0
redis==5.0.1
requests==2.31.0
scikit-learn>=1.3.2
pandas>=2.1.4
numpy>=1.26.0
matplotlib==3.8.2
seaborn==0.13.0
reportlab==4.0.7
openpyxl==3.1.2
cryptography==41.0.7
PyJWT==2.8.0
django-filter==23.3
django-ratelimit==4.1.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url>=2.1